refactor: removed dead code, moved plots draw functions to a separete file
